Move CalledMethods Checker test to the correct location.
