+++ checker/src/test/java/org/checkerframework/checker/test/junit/CalledMethodsDisableframeworksTest.java
@@ -1,17 +1,16 @@
-package tests;
+package org.checkerframework.checker.test.junit;
 
 import java.io.File;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 import org.checkerframework.checker.calledmethods.CalledMethodsChecker;
-import org.checkerframework.framework.test.*;
+import org.checkerframework.framework.test.CheckerFrameworkPerDirectoryTest;
 import org.junit.runners.Parameterized.Parameters;
 
-/** Test case for Called Methods Checker's AutoValue support. */
-public class CalledMethodsAutoValueTest extends CheckerFrameworkPerDirectoryTest {
+public class CalledMethodsDisableframeworksTest extends CheckerFrameworkPerDirectoryTest {
 
-    public CalledMethodsAutoValueTest(List<File> testFiles) {
+    public CalledMethodsDisableframeworksTest(List<File> testFiles) {
         super(
                 testFiles,
                 Arrays.asList(
@@ -21,14 +20,18 @@ public class CalledMethodsAutoValueTest extends CheckerFrameworkPerDirectoryTest
                         "com.google.auto.value.processor.AutoValueBuilderProcessor",
                         "com.google.auto.value.processor.AutoValueProcessor",
                         CalledMethodsChecker.class.getName()),
-                "calledmethods-autovalue",
+                "calledmethods-disableframeworks",
                 Collections.emptyList(),
                 "-Anomsgtext",
+                "-AdisableBuilderFrameworkSupports=autovalue,lombok",
+                // The next option is so that we can run the usevaluechecker tests under this
+                // configuration.
+                "-ACalledMethodsChecker_useValueChecker",
                 "-nowarn");
     }
 
     @Parameters
     public static String[] getTestDirs() {
-        return new String[] {"calledmethods-autovalue"};
+        return new String[] {"calledmethods-disableframeworks", "calledmethods-usevaluechecker"};
     }
 }
